// pharmacy-ui/src/app/services/models/order-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PurchaseProductRequest } from '../models/purchase-product-request';
import { Users } from '../models/users';
export interface OrderRequest {
  customers: Users;
  paymentMethod: 'MPESA' | 'BANK' | 'BITCOIN' | 'VISA';
  products: Array<PurchaseProductRequest>;
  reference: string;
  totalAmount?: number;
}